+++ DependencyInjection/Configuration.php
@@ -19,6 +19,16 @@ class Configuration implements ConfigurationInterface
         $treeBuilder = new TreeBuilder();
         $rootNode = $treeBuilder->root('c975_l_exception_checker');
 
+        $rootNode
+            ->children()
+                ->scalarNode('roleNeeded')
+                    ->defaultValue('ROLE_ADMIN')
+                ->end()
+                ->scalarNode('redirectExcluded')
+                ->end()
+            ->end()
+        ;
+
         return $treeBuilder;
     }
 }