+++ DependencyInjection/c975LExceptionCheckerExtension.php
@@ -27,5 +27,11 @@ class c975LExceptionCheckerExtension extends Extension
             new FileLocator(__DIR__.'/../Resources/config')
         );
         $loader->load('services.yml');
+
+        $configuration = new Configuration();
+        $processedConfig = $this->processConfiguration($configuration, $configs);
+
+        $container->setParameter('c975_l_exception_checker.roleNeeded', $processedConfig['roleNeeded']);
+        $container->setParameter('c975_l_exception_checker.redirectExcluded', $processedConfig['redirectExcluded']);
     }
 }